Move admin session minting into UserController
