+++ services/web/app/src/router.mjs
@@ -22,7 +22,6 @@ import UserInfoController from './Features/User/UserInfoController.js'
 import UserController from './Features/User/UserController.js'
 import UserEmailsController from './Features/User/UserEmailsController.js'
 import UserPagesController from './Features/User/UserPagesController.js'
-import AdminSessionController from './Features/Admin/AdminSessionController.js'
 import TutorialController from './Features/Tutorial/TutorialController.js'
 import DocumentController from './Features/Documents/DocumentController.js'
 import CompileManager from './Features/Compile/CompileManager.js'
@@ -500,7 +499,7 @@ async function initialize(webRouter, privateApiRouter, publicApiRouter) {
   privateApiRouter.post(
     '/internal/admin/session',
     AuthenticationController.requirePrivateApiAuth(),
-    AdminSessionController.createSession
+    UserController.createSession
   )
 
   webRouter.get(
